Change to be exclusive of boundaries to match FilterByTime
Previously the behaviour was inclusive of boundaries, so we would end up
with adjacent log values being included in the filtered result. This
commit changes behaviour so they are exclusive of to match the behaviour
of the filtering done by FilterByTime.

I believe there was an error in one of the FilterEvents unit tests so I have
updated the result and explained the calculation in the comment. I think
this was to do with the fact that the time-series-property index was
being decremented in each loop to account for an "overlap" which meant
that some of the values were being double-counted. My best guess is that
this was being done to allow for the fact that the time-series values
may need to be used multiple times if there are multiple groups. Instead
I've accounted for this by keeping track of the time-series index
separately for each group.

Re #27556