Add an option to exclude rows in the GUI
This commit adds the facility to be able to exclude rows from processing in the reflectometry GUI and to annotate the reason. This is done by adding a property for the exclude reason to the underlying algorithm. Although this doesn't make so much sense at the algorithm level, this algorithm is only really used for the GUI so I think this is ok. This approach gives the benefit that we do not need to add additional complication to the table in the GUI, which is already quite large. This is also much easier to implement. The row is simply skipped if the reason is set. The row is also shaded grey and the reason set as a tooltip (because the Options column is not always visible if the window is not wide enough).