Remove tests for unsupported sum in Q functionality
We no longer support sum in Q for a point detector (this was never required anyway but now it throws, so the test has been updated accordingly).
We no longer support sum in Q for multiple regions of interest (groups). This is not currently used and I'm not sure it's likely to be. It could always be added in future but is beyond scope of the current functionality so it is ok to remove this for now. This means we can easily use the new sum in Q algorithm and fix the workflow algorithm history, which is much more important.

Re #23012